refactor(event): moved settings from entrypoint/fastapi to root
