+++ apps/events/src/events/settings.py
